+++ core/src/test/scala/commbank/coppersmith/FeatureBuilderSpec.scala
@@ -16,6 +16,7 @@ package commbank.coppersmith
 
 import org.joda.time.DateTime
 import org.scalacheck.Prop.forAll
+import org.scalacheck.Gen
 
 import org.specs2._
 import org.specs2.matcher.Matcher
@@ -28,7 +29,7 @@ import scalaz.syntax.std.list.ToListOpsFromList
 import Feature._, Value._, Type._
 import FeatureBuilderSource.fromFS
 import Arbitraries._
-import test.thrift.Customer
+import commbank.coppersmith.test.thrift.{Account, Customer}
 
 object SelectFeatureSetSpec extends Specification with ScalaCheck { def is = s2"""
   SelectFeatureSet - Test an example set of features based on selecting fields
@@ -68,11 +69,11 @@ object SelectFeatureSetSpec extends Specification with ScalaCheck { def is = s2"
     import CustomerFeatureSet.namespace
 
     metadata must_== List(
-      Metadata[Customer, Integral](namespace, "age",       "Age",            Ordinal),
-      Metadata[Customer, Integral](namespace, "tallAge",   "Tall Age",       Continuous),
-      Metadata[Customer, Decimal] (namespace, "oldHeight", "Old Height",     Continuous),
-      Metadata[Customer, Decimal] (namespace, "credit",    "Credit Score",   Continuous),
-      Metadata[Customer, Decimal] (namespace, "altCredit", "Alternate Impl", Continuous)
+      Metadata[Customer, Integral]     (namespace, "age",       "Age",            Ordinal),
+      Metadata[Customer, Integral]     (namespace, "tallAge",   "Tall Age",       Continuous),
+      Metadata[Customer, FloatingPoint](namespace, "oldHeight", "Old Height",     Continuous),
+      Metadata[Customer, FloatingPoint](namespace, "credit",    "Credit Score",   Continuous),
+      Metadata[Customer, FloatingPoint](namespace, "altCredit", "Alternate Impl", Continuous)
     )
   }
 
@@ -84,11 +85,11 @@ object SelectFeatureSetSpec extends Specification with ScalaCheck { def is = s2"
     val expectCredit    = c.credit.isDefined
 
     featureValues must_== List(
-                        Some(FeatureValue[Integral](c.id, "age",       c.age)),
-      expectTallAge.option(  FeatureValue[Integral](c.id, "tallAge",   c.age)),
-      expectOldHieght.option(FeatureValue[Decimal] (c.id, "oldHeight", c.height)),
-      expectCredit.option(   FeatureValue[Decimal] (c.id, "credit",    c.credit)),
-      expectCredit.option(   FeatureValue[Decimal] (c.id, "altCredit", c.credit))
+                        Some(FeatureValue[Integral]     (c.id, "age",       c.age)),
+      expectTallAge.option(  FeatureValue[Integral]     (c.id, "tallAge",   c.age)),
+      expectOldHieght.option(FeatureValue[FloatingPoint](c.id, "oldHeight", c.height)),
+      expectCredit.option(   FeatureValue[FloatingPoint](c.id, "credit",    c.credit)),
+      expectCredit.option(   FeatureValue[FloatingPoint](c.id, "altCredit", c.credit))
     ).flatten
   }}
 }
@@ -99,6 +100,8 @@ object AggregationFeatureSetSpec extends Specification with ScalaCheck { def is 
   An example feature set
     must generate expected metadata       $generateMetadata
     must generate expected feature values $generateFeatureValues
+  The avg aggregator
+    must generate expected values         $generateAvgValues
 """
 
   object CustomerFeatureSet extends AggregationFeatureSet[Customer] {
@@ -112,16 +115,21 @@ object AggregationFeatureSetSpec extends Specification with ScalaCheck { def is 
 
     type CAF = AggregationFeature[Customer, Customer, _, Value]
 
-    val sizeF:    CAF = select(size)                      .asFeature(Discrete,    "size",  "Agg feature")
-    val bigSizeF: CAF = select(size).having(_ > 5)        .asFeature(Discrete,    "sizeB", "Agg feature")
-    val countF:   CAF = select(count(where = _.age >= 18)).asFeature(Continuous,  "count", "Agg feature")
-    val sumF:     CAF = select(sum(_.height))             .asFeature(Continuous,  "sum",   "Agg feature")
-    val maxF:     CAF = select(max(_.age))                .asFeature(Continuous,  "max",   "Agg feature")
-    val minF:     CAF = select(min(_.height))             .asFeature(Continuous,  "min",   "Agg feature")
-    val maxByF:   CAF = select(maxBy(_.id)(_.height))     .asFeature(Continuous,  "maxBy", "Agg feature")
-    val minByF:   CAF = select(minBy(_.id)(_.age))        .asFeature(Continuous,  "minBy", "Agg feature")
-    val avgF:     CAF = select(avg(_.age.toDouble))       .asFeature(Continuous,  "avg",   "Agg feature")
-    val ucbF:     CAF = select(uniqueCountBy(_.age % 10)) .asFeature(Continuous,  "ucb",   "Agg feature")
+    def bigD(f: Customer => Double): (Customer => BigDecimal) = c => BigDecimal(f(c))
+
+    val sizeF:    CAF = select(size)                      .asFeature(Discrete,   "size",    "Agg feature")
+    val bigSizeF: CAF = select(size).having(_ > 5)        .asFeature(Discrete,   "sizeB",   "Agg feature")
+    val countF:   CAF = select(count(where = _.age >= 18)).asFeature(Continuous, "count",   "Agg feature")
+    val sumF:     CAF = select(sum(_.height))             .asFeature(Continuous, "sum",     "Agg feature")
+    val maxF:     CAF = select(max(_.age))                .asFeature(Continuous, "max",     "Agg feature")
+    val minF:     CAF = select(min(_.height))             .asFeature(Continuous, "min",     "Agg feature")
+    val maxByF:   CAF = select(maxBy(_.id)(_.height))     .asFeature(Continuous, "maxBy",   "Agg feature")
+    val minByF:   CAF = select(minBy(_.id)(_.age))        .asFeature(Continuous, "minBy",   "Agg feature")
+    val avgF:     CAF = select(avg(_.age.toDouble))       .asFeature(Continuous, "avg",     "Agg feature")
+    val sumBigDF: CAF = select(sum(bigD(_.height)))       .asFeature(Continuous, "sumBigD", "Agg feature")
+    val maxBigDF: CAF = select(max(bigD(_.age)))          .asFeature(Continuous, "maxBigD", "Agg feature")
+    val avgBigDF: CAF = select(avgBigDec(_.age))          .asFeature(Continuous, "avgBigD", "Agg feature")
+    val ucbF:     CAF = select(uniqueCountBy(_.age % 10)) .asFeature(Continuous, "ucb",     "Agg feature")
 
     import com.twitter.algebird.Aggregator
 
@@ -130,7 +138,22 @@ object AggregationFeatureSetSpec extends Specification with ScalaCheck { def is 
         case (c, Some(credit)) => credit
       }.select(Aggregator.min[Double]).asFeature(Continuous, "collect", "Agg feature")
 
-    def aggregationFeatures = List(sizeF, bigSizeF, countF, sumF, maxF, minF, maxByF, minByF, avgF, ucbF, collectF)
+    def aggregationFeatures = List(
+      sizeF,
+      bigSizeF,
+      countF,
+      sumF,
+      maxF,
+      minF,
+      maxByF,
+      minByF,
+      avgF,
+      sumBigDF,
+      maxBigDF,
+      avgBigDF,
+      ucbF,
+      collectF
+    )
   }
 
   def generateMetadata = {
@@ -138,17 +161,20 @@ object AggregationFeatureSetSpec extends Specification with ScalaCheck { def is 
     import CustomerFeatureSet.namespace
 
     metadata must_== List(
-      Metadata[(EntityId, Iterable[Customer]), Integral](namespace, "size",    "Agg feature",  Discrete),
-      Metadata[(EntityId, Iterable[Customer]), Integral](namespace, "sizeB",   "Agg feature",  Discrete),
-      Metadata[(EntityId, Iterable[Customer]), Integral](namespace, "count",   "Agg feature",  Continuous),
-      Metadata[(EntityId, Iterable[Customer]), Decimal] (namespace, "sum",     "Agg feature",  Continuous),
-      Metadata[(EntityId, Iterable[Customer]), Integral](namespace, "max",     "Agg feature",  Continuous),
-      Metadata[(EntityId, Iterable[Customer]), Decimal] (namespace, "min",     "Agg feature",  Continuous),
-      Metadata[(EntityId, Iterable[Customer]), Decimal] (namespace, "maxBy",   "Agg feature",  Continuous),
-      Metadata[(EntityId, Iterable[Customer]), Integral](namespace, "minBy",   "Agg feature",  Continuous),
-      Metadata[(EntityId, Iterable[Customer]), Decimal] (namespace, "avg",     "Agg feature",  Continuous),
-      Metadata[(EntityId, Iterable[Customer]), Integral](namespace, "ucb",     "Agg feature",  Continuous),
-      Metadata[(EntityId, Iterable[Customer]), Decimal] (namespace, "collect", "Agg feature",  Continuous)
+      Metadata[(EntityId, Iterable[Customer]), Integral]     (namespace, "size",    "Agg feature", Discrete),
+      Metadata[(EntityId, Iterable[Customer]), Integral]     (namespace, "sizeB",   "Agg feature", Discrete),
+      Metadata[(EntityId, Iterable[Customer]), Integral]     (namespace, "count",   "Agg feature", Continuous),
+      Metadata[(EntityId, Iterable[Customer]), FloatingPoint](namespace, "sum",     "Agg feature", Continuous),
+      Metadata[(EntityId, Iterable[Customer]), Integral]     (namespace, "max",     "Agg feature", Continuous),
+      Metadata[(EntityId, Iterable[Customer]), FloatingPoint](namespace, "min",     "Agg feature", Continuous),
+      Metadata[(EntityId, Iterable[Customer]), FloatingPoint](namespace, "maxBy",   "Agg feature", Continuous),
+      Metadata[(EntityId, Iterable[Customer]), Integral]     (namespace, "minBy",   "Agg feature", Continuous),
+      Metadata[(EntityId, Iterable[Customer]), FloatingPoint](namespace, "avg",     "Agg feature", Continuous),
+      Metadata[(EntityId, Iterable[Customer]), Decimal]      (namespace, "sumBigD", "Agg feature", Continuous),
+      Metadata[(EntityId, Iterable[Customer]), Decimal]      (namespace, "maxBigD", "Agg feature", Continuous),
+      Metadata[(EntityId, Iterable[Customer]), Decimal]      (namespace, "avgBigD", "Agg feature", Continuous),
+      Metadata[(EntityId, Iterable[Customer]), Integral]     (namespace, "ucb",     "Agg feature", Continuous),
+      Metadata[(EntityId, Iterable[Customer]), FloatingPoint](namespace, "collect", "Agg feature", Continuous)
     )
   }
 
@@ -168,17 +194,20 @@ object AggregationFeatureSetSpec extends Specification with ScalaCheck { def is 
     val expectBig   = cs.size > 5
 
     val expected = List(
-                  Some((c.id, "size",    cs.size:                         Integral, time)),
-      expectBig.option((c.id, "sizeB",   cs.size:                         Integral, time)),
-                  Some((c.id, "count",   ages.filter(_ >= 18).size:       Integral, time)),
-                  Some((c.id, "sum",     heights.sum:                     Decimal,  time)),
-                  Some((c.id, "max",     ages.max:                        Integral, time)),
-                  Some((c.id, "min",     heights.min:                     Decimal,  time)),
-                  Some((c.id, "maxBy",   maxCById.height:                 Decimal,  time)),
-                  Some((c.id, "minBy",   minCById.age:                    Integral, time)),
-                  Some((c.id, "avg",     (ages.sum / ages.size.toDouble): Decimal,  time)),
-                  Some((c.id, "ucb",     groupedAges.size:                Integral, time)),
-      credit.map(d =>  (c.id, "collect", d:                               Decimal,  time))
+                  Some((c.id, "size",    cs.size:                            Integral,      time)),
+      expectBig.option((c.id, "sizeB",   cs.size:                            Integral,      time)),
+                  Some((c.id, "count",   ages.filter(_ >= 18).size:          Integral,      time)),
+                  Some((c.id, "sum",     heights.sum:                        FloatingPoint, time)),
+                  Some((c.id, "max",     ages.max:                           Integral,      time)),
+                  Some((c.id, "min",     heights.min:                        FloatingPoint, time)),
+                  Some((c.id, "maxBy",   maxCById.height:                    FloatingPoint, time)),
+                  Some((c.id, "minBy",   minCById.age:                       Integral,      time)),
+                  Some((c.id, "avg",     (ages.sum.toDouble / ages.size):    FloatingPoint, time)),
+                  Some((c.id, "sumBigD", heights.map(BigDecimal(_)).sum:     Decimal,       time)),
+                  Some((c.id, "maxBigD", ages.map(BigDecimal(_)).max:        Decimal,       time)),
+                  Some((c.id, "avgBigD", (ages.sum / BigDecimal(ages.size)): Decimal,       time)),
+                  Some((c.id, "ucb",     groupedAges.size:                   Integral,      time)),
+      credit.map(d =>  (c.id, "collect", d:                                  FloatingPoint, time))
     ).flatten
 
     eavtValues.toList must matchEavts(expected)
@@ -189,13 +218,52 @@ object AggregationFeatureSetSpec extends Specification with ScalaCheck { def is 
     expected.contain(_.zip(===, ===, matchValue, ===))
 
   def matchValue(expected: Value): Matcher[Value] = expected match {
-    case Integral(_) | Str(_) | Decimal(None) => be_===(expected)
-    case Decimal(Some(expectedDouble)) =>
+    case Integral(_) | Str(_) | FloatingPoint(None) | Decimal(None) => be_===(expected)
+    case FloatingPoint(Some(expectedDouble)) =>
       beCloseTo(expectedDouble +/- 0.000000000001) ^^ (
         (v: Value) => v match {
-          case Decimal(Some(actualDouble)) => actualDouble
+          case FloatingPoint(Some(actualDouble)) => actualDouble
           case _ => Double.NaN
         }
       )
+    case Decimal(Some(expectedBigDecimal)) =>
+      beCloseTo(expectedBigDecimal +/- BigDecimal("0.0000000000000000001")) ^^ (
+        (v: Value) => v match {
+          case Decimal(Some(actualBigDecimal)) => actualBigDecimal
+          case _ => null
+        }
+      )
+  }
+
+  object AverageFeatureSet extends AggregationFeatureSet[Account] {
+    val namespace                               = "test.namespace"
+    def entity(a: Account)                      = a.id
+    def time(a: Account, ctx: FeatureContext)   = ctx.generationTime.getMillis
+
+    val source  = From[Account]()
+    val builder = source.featureSetBuilder(namespace, entity)
+    val select: FeatureSetBuilder[Account, Account] = builder
+
+    type CAF      = AggregationFeature[Account, Account, _, Value]
+    val avgF: CAF = select(avgBigDec(c => BigDecimal(c.balanceBigDecimal)))
+      .asFeature(Continuous, "avg", "Agg feature")
+
+    def aggregationFeatures = List(avgF)
+  }
+
+  def generateAvgValues = {
+    testAverage(NonEmptyList("1000000000000.02", "0.02"), BigDecimal("500000000000.02"))
+    testAverage(NonEmptyList("1200000000000.01", "1200000000000.03", "0.02"), BigDecimal("800000000000.02"))
+  }
+
+  def testAverage(vals: NonEmptyList[String], expected: BigDecimal) = {
+    val as   = vals.map(v => Gen.resultOf(Account.apply _).sample.get.copy(balanceBigDecimal = v))
+    val time = DateTime.now.getMillis
+
+    val featureValues = AverageFeatureSet.generate((as.head.id, as.list))
+
+    val eavtValues = featureValues.map { fv => fv.asEavt(time) }.toList
+
+    eavtValues.toList must matchEavts(List((as.head.id, "avg", expected: Decimal, time)))
   }
 }